[FF-A] Notification Service Updates (#339)
## Description

Updated the ID and Bitmap values to UINT32s.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?
- [ ] Backport to release branch?

## How This Was Tested

Tested with the FfaPartitionTestApp

## Integration Instructions

N/A

(cherry picked from commit fc8f6fd677f6de60e597b0edadeb9fab2febe544)

+++ FfaFeaturePkg/Library/TestServiceLib/TestServiceLib.c
@@ -43,7 +43,7 @@ TestNotificationHandler (
   /* Set the notification set flag to be a delayed SRI */
   UINT32  Flag = (1 << DELAYED_SRI_BIT_POS);
 
-  UINT16              LogicalId = Request->Arg3;
+  UINT32              LogicalId = Request->Arg3;
   NotificationStatus  Status    = NotificationServiceIdSet (LogicalId, Uuid, Flag);
 
   /* Check for a valid UUID and validate the input parameters */
